All the work I did over holiday in Mexico

+++ db/migrate/20180719203827_add_length_to_sailboats.rb
@@ -1,5 +1,5 @@
 class AddLengthToSailboats < ActiveRecord::Migration[5.2]
   def change
-    add_column :sailboats, :length, :integer
+    add_column :sailboats, :length, :decimal
   end
 end
